chore(refactor): refactor package.json for heroku deploy
