+++ package.json
@@ -4,15 +4,15 @@
   "description": "full stack web messaging application",
   "scripts": {
     "babel-version": "babel --version",
-    "build": "babel src -d lib",
+    "build": "webpack -p",
     "coveralls": "nyc --reporter=lcov --reporter=text-lcov npm test",
     "exec": "nodemon --watch ./server/bin/www --exec babel-node -- server/app",
     "start": "nodemon ./server/bin/www --exec babel-node .server/app",
     "test_": "nyc mocha --compilers js:babel-core/register --timeout 10000",
     "test-dev": "mocha --compilers js:babel-core/register --timeout 10000",
     "test": "nyc --reporter=html --reporter=text mocha --compilers js:babel-register",
-    "start-client": "webpack -p && webpack-dev-server",
-    "start-dev": "webpack-dev-server"
+    "build-client": "webpack -p && webpack-dev-server",
+    "heroku-postbuild": ""
   },
   "repository": {
     "type": "git",
